separating components
