+++ src/js/components/megaMenu/MegaMenu.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import AnimatedNavbar from './AnimatedNavbar';
+
+const MegaMenu = () => {
+    const duration = 300, ease = "easeOutExpo";
+
+    return (
+        <div className="mega-menu-container">
+            <AnimatedNavbar
+                tweenConfig={{
+                    ease: ease,
+                    duration: duration
+                }}
+            />
+        </div>
+    );
+}
+
+export default MegaMenu;
